# Speedup decode by 1.6x (+159 tok/s) by removing costly NaN checks